Fix hanging on close
Due to processes not being terminated and joined the app would hang while closing. The True loop inside the worker function was fixed, the close task definition was fixed, and the close function was fixed by adding the terminate command to close the process. The event close function was also added to the OnClose BaseApp function